// types.ts
export enum SkinType {
  GLASS = 'skin-glass',
  NEON = 'skin-neon',
  MINIMAL = 'skin-minimal',
  GRADIENT = 'skin-gradient',
  MATTE = 'skin-matte',
  PARCHMENT = 'skin-hp-parchment',
  EMERALD = 'skin-hp-emerald',
  GLITCH = 'skin-glitch',
  OCEANIC = 'skin-oceanic',
  ARCADE = 'skin-arcade'
}

export interface LinkButton {
  id: string;
  label: string;
  url: string;
}